+++ packages/varlet-ui/src/popup/index.ts
@@ -1,5 +1,5 @@
 import type { App } from 'vue'
-import Popup from './Popup.vue'
+import Popup from './Popup'
 
 Popup.install = function (app: App) {
   app.component(Popup.name, Popup)