+++ packages/varlet-ui/src/styles/example/index.vue
@@ -1,18 +1,28 @@
 <template>
   <div class="list">
-    <div class="list__item" :class="`var-elevation--${e}`" v-for="e in elevations" :key="e">
+    <div class="list__item" :class="`var-elevation--${e}`" :style="{ background }" v-for="e in elevations" :key="e">
       {{ e }}
     </div>
   </div>
 </template>
 
 <script>
+import dark from '../../themes/dark'
+import { ref } from 'vue'
+import { watchDarkMode } from '@varlet/cli/site/utils'
+
 export default {
   name: 'RippleExample',
   setup() {
     const elevations = Array.from({ length: 25 }).map((_, index) => index)
+    const background = ref(dark)
+
+    watchDarkMode(dark, (themes) => {
+      background.value = themes === 'darkThemes' ? '#303030' : '#fff'
+    })
 
     return {
+      background,
       elevations,
     }
   },
@@ -35,7 +45,6 @@ export default {
     height: 25vw;
     color: var(--site-config-color-sub-text);
     margin: 2vw;
-    background: var(--site-config-color-sub-bar);
     transition: 0.25s background-color;
   }
 }