+++ packages/varlet-ui/src/themes/dark/card.ts
@@ -1,5 +1,5 @@
 export default {
-  '--card-background': '#272727',
+  '--card-background': '#303030',
   '--card-title-color': '#ffffff',
   '--card-subtitle-color': '#aaaaaa',
   '--card-description-color': '#aaaaaa',