+++ packages/varlet-ui/src/themes/dark/datePicker.ts
@@ -1,5 +1,5 @@
 export default {
   '--date-picker-main-color': '#fff',
-  '--date-picker-body-background-color': '#272727',
+  '--date-picker-body-background-color': '#303030',
   '--day-picker-head-item-color': '#aaaaaa',
 }