+++ packages/varlet-ui/src/themes/dark/select.ts
@@ -1,5 +1,5 @@
 export default {
   '--select-select-text-color': '#fff',
   '--select-blur-color': 'rgb(255, 255, 255, .7)',
-  '--select-scroller-background': '#272727',
+  '--select-scroller-background': '#303030',
 }