+++ packages/varlet-ui/src/themes/dark/table.ts
@@ -1,5 +1,5 @@
 export default {
-  '--table-background': '#272727',
+  '--table-background': '#303030',
   '--table-thead-th-text-color': 'rgba(255, 255, 255, 0.6)',
   '--table-tbody-tr-hover-background': '#4c4b4b',
   '--table-thead-border-bottom': 'thin solid rgba(255, 255, 255, 0.12)',