+++ packages/varlet-ui/types/global.d.ts
@@ -26,6 +26,7 @@ declare module 'vue' {
     VarIndexBar: typeof import('@varlet/ui')['_IndexBarComponent']
     VarInput: typeof import('@varlet/ui')['_InputComponent']
     VarLazy: typeof import('@varlet/ui')['_LazyComponent']
+    VarList: typeof import('@varlet/ui')['_ListComponent']
     VarLoading: typeof import('@varlet/ui')['_LoadingComponent']
     VarLocale: typeof import('@varlet/ui')['_LocaleComponent']
     VarMenu: typeof import('@varlet/ui')['_MenuComponent']