+++ packages/varlet-cli/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.5](https://github.com/haoziqaq/varlet/compare/v1.23.4...v1.23.5) (2021-11-21)
+
+
+### Bug Fixes
+
+* **cli:** add module moduleCompatible option in varlet.config.js ([0b1d95f](https://github.com/haoziqaq/varlet/commit/0b1d95fc740a225240ee3f28e06015f4538b2304))
+* **cli:** expose lib folder ([2a8848a](https://github.com/haoziqaq/varlet/commit/2a8848a6f064aa91d56875ba467c46da79589a79))
+* **cli:** fix commonjs compiler ([c5c2b12](https://github.com/haoziqaq/varlet/commit/c5c2b12f163bed130b3c6c0ee0eb805fe4a015f1))
+* **cli:** fix inline css plugin ([3e8b7b4](https://github.com/haoziqaq/varlet/commit/3e8b7b4e5489bc67dd38f8787477f07e2a13e7ac))
+* **cli/compile:** add commonjs bundle ([4a424d4](https://github.com/haoziqaq/varlet/commit/4a424d4c066f7de69cbf561a5bc9279862739637))
+* **cli/compile:** add match commonjs script extname function ([2b5b1bd](https://github.com/haoziqaq/varlet/commit/2b5b1bde5e1c16d4e58ce2a100ae4a1a3cc428ce))
+* **cli/compile:** fix compile module readdir ([9c3ad08](https://github.com/haoziqaq/varlet/commit/9c3ad08410009e50b288f033b8fdbe10a8c01160))
+* **ui/popup,input:** support nuxt ([def81ef](https://github.com/haoziqaq/varlet/commit/def81ef821a8ce679287702316bd98ca363bc7f4))
+
+
+
+
+
+## [1.23.3](https://github.com/haoziqaq/varlet/compare/v1.23.2...v1.23.3) (2021-11-20)
+
+
+### Bug Fixes
+
+* **cli/jest:** add watchAll mode ([84ebb4e](https://github.com/haoziqaq/varlet/commit/84ebb4efe823cb9d7f22995d429b82d8fc3e05c8))
+
+
+
+
+
+# [1.23.0](https://github.com/haoziqaq/varlet/compare/v1.22.10...v1.23.0) (2021-11-19)
+
+
+### Bug Fixes
+
+* **ui,cli:** move dark theme import to example ([579e461](https://github.com/haoziqaq/varlet/commit/579e461a620c2a7e416193b43368b885ec02b862))
+* add assert for example lang ([7ce8bae](https://github.com/haoziqaq/varlet/commit/7ce8baefeaaf0f10d2638cb4927ccbf570b28006))
+* **cli:** bump typescript and commander version ([af893e9](https://github.com/haoziqaq/varlet/commit/af893e970b71ddc46365187c94ecba244513ef5a))
+* **doc:** progress bar should not use unmount function ([0587103](https://github.com/haoziqaq/varlet/commit/05871038e42a50f38fdc2a201107d786431691cd))
+* **site:** modify the width of nav ([af0bfbe](https://github.com/haoziqaq/varlet/commit/af0bfbe8557182fa23083b8a4060625b18132c92))
+* **site:** support bg-color for dark mode of nav button ([b8417fe](https://github.com/haoziqaq/varlet/commit/b8417fe0873b8ecea4bcef21068aa0328ec32f88))
+
+
+### Features
+
+* **all:** support dark mode ([94d9899](https://github.com/haoziqaq/varlet/commit/94d9899cd54ebe1ac864712484d2c6f6f958150a))
+* **cli/generators:** support dark mode ([4795af8](https://github.com/haoziqaq/varlet/commit/4795af8c25cc155eb1fe7cd8d8a6b23b45379989))
+* **menu:** add dark-mode ([e2807f6](https://github.com/haoziqaq/varlet/commit/e2807f6f45ee3331acdd3b9ba4180b16096f0114))
+
+
+
+
+
 ## [1.22.10](https://github.com/haoziqaq/varlet/compare/v1.22.9...v1.22.10) (2021-11-03)